+++ network/src/validator_network/network_builder.rs
@@ -19,11 +19,11 @@ use crate::{
     },
     protocols::{
         discovery::{self, Discovery, PeerInfo},
-        health_checker::HealthChecker,
+        health_checker::{self, HealthChecker},
         identity::Identity,
     },
     transport::*,
-    validator_network, ProtocolId,
+    ProtocolId,
 };
 use channel::{self, libra_channel, message_queues::QueueStyle};
 use futures::stream::StreamExt;
@@ -411,8 +411,7 @@ impl NetworkBuilder {
 
     pub fn add_connection_monitoring(&mut self) -> &mut Self {
         // Initialize and start HealthChecker.
-        let (hc_network_tx, hc_network_rx) =
-            validator_network::health_checker::add_to_network(self);
+        let (hc_network_tx, hc_network_rx) = health_checker::add_to_network(self);
         let ping_interval_ms = self.ping_interval_ms;
         let ping_timeout_ms = self.ping_timeout_ms;
         let ping_failures_tolerated = self.ping_failures_tolerated;
